Json package move to test from main in eureka services
