+++ eureka-services/src/test/java/edu/emory/cci/aiw/cvrg/eureka/services/json/AttributeMixin.java
